[gst-aml-plugins] Add set_video_axis_scaled()
For 4k TV, because we do scaling up in freescale,
wpe-webkit is not aware about this.

It sill think res of full screen is 1080p, and sets
video axis to (0, 0, 1920, 1080).

Add set_video_axis_scaled() to gst-aml-plugins, so
that we can compute scaling factor and adjust video
axis according to it

Signed-off-by: Mingyen Hung <mingyen.hung@amlogic.com>