media_modules: skip size check for dmabuf on output port [1/1]
PD#SWPL-25323

Problem:
When dmabuf is used on output port, the size of it is decided by the
allocator. But when REQBUFS is called to setup the vb2 queue, the
minimium size of plane is returned by queue_setup(). It will lead to a
sanity check failure in __qbuf_dmabuf().

Solution:
When dmabuf mode is used on output port. queue_setup() will return size
0 for it. And vb2 will use the size of the dmabuf itself.

Verify:
U212 + Cobalt + WV

Change-Id: I8feae637e00203b7b2ecf1dfc1535c9bbe99d3ae
Signed-off-by: Song Zhao <song.zhao@amlogic.com>